(PC-17816) refactor(profile): move Subtitle to components + add stories
