+++ src/features/profile/components/Header/CreditHeader/CreditHeader.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { DomainsCredit } from 'api/gen/api'
-import { Subtitle } from 'features/profile/atoms/Subtitle'
+import { Subtitle } from 'features/profile/components/Subtitle/Subtitle'
 import { BeneficiaryCeilings } from 'features/profile/components/BeneficiaryCeilings/BeneficiaryCeilings'
 import { CreditExplanation } from 'features/profile/components/CreditExplanation/CreditExplanation'
 import { CreditInfo } from 'features/profile/components/CreditInfo/CreditInfo'
