+++ src/features/profile/components/Header/NonBeneficiaryHeader/NonBeneficiaryHeader.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components/native'
 import { useAuthContext } from 'features/auth/AuthContext'
 import { useBeneficiaryValidationNavigation } from 'features/auth/signup/useBeneficiaryValidationNavigation'
 import { useNextSubscriptionStep } from 'features/auth/signup/useNextSubscriptionStep'
-import { Subtitle } from 'features/profile/atoms/Subtitle'
+import { Subtitle } from 'features/profile/components/Subtitle/Subtitle'
 import { IdentityCheckPendingBadge } from 'features/profile/components/Badges/IdentityCheckPendingBadge'
 import { SubscriptionMessageBadge } from 'features/profile/components/Badges/SubscriptionMessageBadge'
 import { YoungerBadge } from 'features/profile/components/Badges/YoungerBadge'
